refactor(pswap): move round-build impl + PswapChainNoteUpdate into lineage.rs
Addresses review: the per-round classification methods (build_round_update,
build_*_round, apply_round_in_memory) and the PswapChainNoteUpdate DTO now live
in lineage.rs alongside the PswapLineageRecord constructors/accessors, instead
of a loose impl block in discovery.rs. discovery.rs is now purely the
correlator. Dependency direction is natural (observer/discovery -> lineage).
The round-build unit tests move with them; require_original was inlined.
